requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Signal processing
PyWavelets>=1.4.1

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Evaluation
scikit-learn>=1.3.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0

# Training utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.14.0

# Data download
kaggle>=1.5.16

# Optional: faster spectrogram computation
librosa>=0.10.1
